Cli-Download-Rom/requirements.txt
requests
tqdm
PyYAML
python-dotenv
unidecode
colorama